Regression/Comparing-Regression-Models/requirements.txt
pandas 
numpy 
matplotlib.pyplot 
sklearn
